Reorganize Python Tests
PR feedback indicated that the type-overloaded Python tests were not
quite clear in their intent. These now appear more like the TypeScript
tests: one with cases passed as input and one for the batch version of
the route.

This also switches the `terminology` module's `TypedDict` import to use
`typing_extensions` below 3.11, as `NotRequired` was only made available
in PEP 655.