Remove Base URL, Headers Configuration
It's incredibly uncommon to need to change the base URL for Orchestrate.
Most users will, appropriately, just use api.careevolutionapi.com and
require no additional headers that are not handled in each of the
implemented methods.

As a result, this reduces the configuration in constructors. For Python,
creating a client is now just

```python
api = OrchestrateApi("<key>")
```

and TypeScript:

```typescript
const api = new OrchestrateApi({apiKey: "<key>"});
```

It is still possible to use environment variables to change the base URL
and additional default headers.