Fix some buildy things
Key insight was that in order for us to get expected behavior in editor (like higlighting import errors), we needed to have type: "module" set in the package.json.

Then we can build esm with expected parameters (module: nodeNext).

This means we needed to invert the build-esm-package.mjs script to add the package.json with type:"commonjs" set in the cjs export instead