Report Operational Outcomes in Responses
It was previously very difficult, particularly in Python, to determine
why an operation that returned a 400 status rejected the payload. The
detailed reasons are often reported in OperationOutcomes in the response
body, but these were being swallowed during exception handling.

This adds an exception layer to the SDK, exposing those classes, and
also provides better exposure for those outcomes.