Report Operational Outcomes in Responses (#104)
## Description of Changes

It was previously very difficult, particularly in Python, to determine
why an operation that returned a 400 status rejected the payload. The
detailed reasons are often reported in OperationOutcomes in the response
body, but these were being swallowed during exception handling.

This adds an exception layer to the SDK, exposing those classes, and
also provides better exposure for those outcomes.

## Security

**REMINDER: All file contents are public.**

- [x] I have ensured no secure credentials or sensitive information
remain in code, metadata, comments, etc. Of particular note: No
temporary testing changes committed such as API base URLs, access
tokens, print/log statements, etc.
- [x] My changes do not introduce any security risks, or any such risks
have been properly mitigated.

This changes some output from the SDK, but these outcomes were
previously already available to the client: this just wraps them in
clear exceptions that can be handled and inspected.

There are new JSON and XML parsers to handle this, but all of the
inspection is performed on responses from the server.

## Reviewers

- [x] I have assigned the appropriate reviewer(s).