Send Bytes in Python (#105)
## Description of Changes

The requests library seems to be handling the `data` parameter it
accepts in `POST`s in such a way to cause differently-encoded strings to
fail when passing to Orchestrate. This PR changes the behavior of the
Python library so that it always passes bytes to the `requests` library.
None of the signatures have changed.

## Security

**REMINDER: All file contents are public.**

- [x] I have ensured no secure credentials or sensitive information
remain in code, metadata, comments, etc. Of particular note: No
temporary testing changes committed such as API base URLs, access
tokens, print/log statements, etc.
- [x] My changes do not introduce any security risks, or any such risks
have been properly mitigated.

This only changes the encoding of values passed to Orchestrate.

## Reviewers

- [x] I have assigned the appropriate reviewer(s).